Test de deplacement avec un GamePanel

+++ src/main/java/modele/Ennemi1.java
@@ -2,12 +2,14 @@ package modele;
 
 public class Ennemi1 extends Personnage {
 
-    public Ennemi1(float pointVie, int pointAttaque, Environnement environnement, int positionX, int positionY) {
+    public Ennemi1(int pointVie, int pointAttaque, Environnement environnement, int positionX, int positionY) {
         super(pointVie, pointAttaque, environnement, positionX, positionY, 2);
     }
 
-    public Ennemi1(float pointVie, int pointAttaque, Environnement environnement) {
+    public Ennemi1(int pointVie, int pointAttaque, Environnement environnement) {
         super(pointVie, pointAttaque, environnement, 2);
     }
+
+
 }
